pyproject.toml
[project]
name = "buffotte"
version = "0.13.0"
description = "Buffotte project"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "python-dotenv",
    "pymysql",
    "bcrypt",
    "fastapi",
    "uvicorn",
    "pandas",
    "scikit-learn",
    "lightgbm",
    "openai",
    "pytz",
    "cryptography",
    "httpx",
    "playwright",
    "markdown",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "crawler", "db", "llm", "models"]
